// legacy/apps/web/src/types.ts
export interface CodocListItem {
  id: string;
  path: string;
  nodeState: string;
  meta: {
    title?: string;
    description?: string;
    tags?: string[];
  };
}